@@ -76,79 +76,79 @@
     "url": "https://github.com/tinytacoteam/zazu/issues"
   },
   "dependencies": {
-    "auto-launch": "^5.0.5",
+    "auto-launch": "^5.0.6",
     "cuid": "^2.1.6",
-    "decompress": "^4.2.0",
-    "fs-jetpack": "^2.2.2",
-    "interweave": "^11.1.0",
-    "mkdirp": "^0.5.1",
-    "mousetrap": "^1.6.3",
+    "decompress": "^4.2.1",
+    "fs-jetpack": "^5.1.0",
+    "interweave": "^13.1.0",
+    "mkdirp": "^3.0.1",
+    "mousetrap": "^1.6.5",
     "nestdb": "^2.0.0",
-    "node-notifier": "^5.4.0",
-    "npm": "^6.9.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "node-notifier": "^10.0.1",
+    "npm": "^10.2.3",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "relaxed-json": "^1.0.3",
     "request": "^2.88.0",
-    "semver": "^6.1.1",
-    "winston": "^3.2.1",
-    "winston-daily-rotate-file": "^3.9.0",
-    "winston-transport": "^4.3.0"
+    "semver": "^7.5.4",
+    "winston": "^3.11.0",
+    "winston-daily-rotate-file": "^4.7.1",
+    "winston-transport": "^4.6.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.23.2",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-react-constant-elements": "^7.22.5",
+    "@babel/plugin-transform-react-inline-elements": "^7.22.5",
+    "@babel/preset-env": "^7.23.2",
+    "@babel/preset-react": "^7.22.15",
+    "@babel/register": "^7.22.15",
+    "@babel/runtime": "^7.23.2",
     "asar": "^2.0.1",
     "babel-eslint": "^10.0.2",
-    "babel-loader": "^8.0.6",
-    "@babel/core": "^7.4.5",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-react-constant-elements": "^7.2.0",
-    "@babel/plugin-transform-react-inline-elements": "^7.2.0",
-    "@babel/preset-env": "^7.4.5",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.4.4",
-    "@babel/runtime": "^7.4.5",
-    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "babel-loader": "^9.1.3",
+    "babel-plugin-closure-elimination": "^1.3.2",
     "babel-plugin-dynamic-import-node-babel-7": "^2.0.7",
-    "babel-plugin-closure-elimination": "^1.3.0",
-    "babel-plugin-macros": "^2.6.1",
-    "chai": "^4.2.0",
-    "cheerio": "^1.0.0-rc.3",
-    "concurrently": "^4.1.0",
-    "cross-env": "^5.2.0",
+    "babel-plugin-macros": "^3.1.0",
+    "chai": "^4.3.10",
+    "cheerio": "^1.0.0-rc.12",
+    "concurrently": "^8.2.2",
+    "cross-env": "^7.0.3",
+    "css-minimizer-webpack-plugin": "^5.0.1",
     "cucumber": "^5.1.0",
-    "electron": "^5.0.4",
-    "electron-builder": "^20.43.0",
-    "electron-mocha": "^8.0.2",
-    "eslint": "^5.16.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-html": "^5.0.5",
-    "eslint-plugin-import": "^2.17.3",
-    "eslint-plugin-node": "^9.1.0",
-    "eslint-plugin-promise": "4.1.1",
-    "eslint-plugin-react": "^7.13.0",
-    "eslint-plugin-standard": "4.0.0",
-    "fs-jetpack": "^2.2.2",
-    "gh-pages": "^2.0.1",
+    "electron": "^27.0.3",
+    "electron-builder": "^24.6.4",
+    "electron-mocha": "^12.1.0",
+    "eslint": "^8.53.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-html": "^7.1.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-standard": "4.1.0",
+    "fs-jetpack": "^5.1.0",
+    "gh-pages": "^6.0.0",
     "gulp": "^4.0.2",
     "gulp-util": "^3.0.8",
-    "husky": "^2.4.1",
+    "husky": "^8.0.3",
     "is-travis": "^2.0.0",
-    "mini-css-extract-plugin": "^0.7.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "mini-css-extract-plugin": "^2.7.6",
     "q": "^1.5.1",
-    "sinon": "^7.3.2",
-    "sinon-chai": "^3.3.0",
-    "spectron": "^5.0.0",
-    "terser-webpack-plugin": "^1.3.0",
-    "webpack": "^4.34.0",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.4",
-    "webpack-merge": "^4.2.1"
+    "sinon": "^17.0.1",
+    "sinon-chai": "^3.7.0",
+    "spectron": "^19.0.0",
+    "terser-webpack-plugin": "^5.3.9",
+    "webpack": "^5.89.0",
+    "webpack-bundle-analyzer": "^4.9.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-merge": "^5.10.0"
   },
   "optionalDependencies": {
-    "appdmg": "^0.6.0",
-    "robotjs": "^0.5.1",
-    "rcedit": "^2.0.0"
+    "appdmg": "^0.6.6",
+    "rcedit": "^4.0.0",
+    "robotjs": "^0.6.0"
   }
 }
